refactor: split markdown::query into per-kind submodules
Each query function now owns its file under modules/libs/src/markdown/query/:

- common.rs   - shared helpers (options, sourcepos→byte conversion,
                code-region detection, wikilink-url parsing)
- images.rs   - images()
- wikilinks.rs - wikilinks() + WikiLinkRef + From<&WikiLinkRef> for ScrapKey
- embeds.rs   - embeds() + EmbedRef
- tags.rs     - tags() + TagOccurrence
- task_items.rs - task_items() + TaskItem + TaskStatus
- section.rs  - section() + GFM slug helper

query.rs becomes the module index with re-exports so the public
surface (`scraps_libs::markdown::query::*`) is unchanged.
